MASSIVE UPDATES - mainly added threshold
Implemented a threshold features that the user inputs (need to make sure
they don't input 0 (?)) Also added a function that will calculate the
changes to the weights/bias so I can use that to determine if the change
will be within the threshold. Updated some of the helper functions to
InitializeIT and testIt(). Renamed a couple functions, and implemented a
function errorDiff to get threshold. I think that's it, other than
deleting old print statements/adding new ones